+++ clang-tools-extra/clang-tidy/embedded/MissingStaticCheck.h
@@ -1,0 +1,30 @@
+//===--- MissingStaticCheck.h - clang-tidy ----------------------*- C++ -*-===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_EMBEDDED_MISSINGSTATICCHECK_H
+#define LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_EMBEDDED_MISSINGSTATICCHECK_H
+
+#include "../ClangTidyCheck.h"
+
+namespace clang::tidy::embedded {
+
+/// Detects inline function declarations that lack the static storage qualifier.
+///
+/// For the user-facing documentation see:
+/// http://clang.llvm.org/extra/clang-tidy/checks/embedded/missing-static.html
+class MissingStaticCheck : public ClangTidyCheck {
+public:
+  MissingStaticCheck(StringRef Name, ClangTidyContext *Context)
+      : ClangTidyCheck(Name, Context) {}
+  void registerMatchers(ast_matchers::MatchFinder *Finder) override;
+  void check(const ast_matchers::MatchFinder::MatchResult &Result) override;
+};
+
+} // namespace clang::tidy::embedded
+
+#endif // LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_EMBEDDED_MISSINGSTATICCHECK_H